added README and search

+++ app/src/main/java/com/java/chenkaiwen/News/Keywords.java
@@ -8,7 +8,7 @@ public class Keywords {
     public String word;
 
     public Keywords(JSONObject o) throws JSONException {
-        this.score = o.getInt("score");
+        this.score = o.getDouble("score");
         this.word = o.getString("word");
     }
 }
